Moved itsNewUsername and itsNewEmail to proper location
